+++ sakura_core/types/Type_Pascal.cpp
@@ -116,5 +116,5 @@ const wchar_t* g_ppszKeywordsPASCAL[] = {
 	L"protected",
 	L"override"
 };
-int g_nKeywordsPASCAL = _countof(g_ppszKeywordsPASCAL);
+size_t g_nKeywordsPASCAL = _countof(g_ppszKeywordsPASCAL);
